+++ docs/docs-components/SandpackExample.js
@@ -123,7 +123,7 @@ export default function SandpackExample({
 
   const { devExampleMode } = useAppContext();
 
-  return process.env.NODE_ENV === 'production' || devExampleMode === 'sandpack' ? (
+  return devExampleMode === 'default' ? (
     <SandpackProvider
       // Based on https://github.com/codesandbox/sandpack/blob/53811bb4fdfb66ea95b9881ff18c93307f12ce0d/sandpack-react/src/presets/Sandpack.tsx#L67
       template="react"
@@ -174,7 +174,7 @@ export default function SandpackExample({
             }
           : {}),
         '/App.js': {
-          code,
+          code: devExampleMode === 'default' ? code : '',
         },
       }}
       theme="dark"